win32

+++ src/fwbuilder/inet_net_pton.c
@@ -23,7 +23,7 @@
 
 #include <stdio.h>
 #include <string.h>
-
+#include <assert.h>
 
 static int  inet_net_pton_ipv4(const char *src, u_char *dst);
 static int  inet_cidr_pton_ipv4(const char *src, u_char *dst, size_t size);
@@ -49,8 +49,7 @@ static int  inet_cidr_pton_ipv6(const char *src, u_char *dst, size_t size);
  *  the names to reflect their current use.
  *
  */
-int
-inet_net_pton(int af, const char *src, void *dst, size_t size)
+int inet_net_pton(int af, const char *src, void *dst, int size)
 {
     switch (af)
     {
